PP-157 Restructure controllers as per frontend pattern Exploit node.js 5.x syntax features to clean things up.
