+++ app/controllers/root_controller.js
@@ -1,0 +1,5 @@
+module.exports.bindRoutesTo = (app) => {
+  app.get('/', (req, res) => {
+    res.redirect(303, '/proceed');
+  });
+}
